Add Monocypher
Note: as the author of Monocypher, I can reasonably speculate this will
not be fixed.  I may provide a higher-level interface that doesn't
exhibit the problem, but the vulnerable API will likely not go away.

The main reason is because Monocypher is meant to be used many different
settings, included weak embedded machines where a single signature takes
up to a minute, and recomputing the public key on the fly is not an
option.  On the other hand, other users may wish to pack their data as
much as possible, and generate the public key on the fly to compensate.

Right now I don't see how I can satisfy both use cases without the
current API.  Strictly speaking even NaCl can do it, but few users are
wicked enough to conserve only the first 32 bytes of the private key and
rebuilding the public half right next to it on the fly.